% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{add_coord}
\alias{add_coord}
\title{Add gene coordinates to homolog data}
\usage{
add_coord(homolog_df, genes_gr, symbol_colunm_name)
}
\arguments{
\item{homolog_df}{A data frame with two columns which contains the homolog
pairs.}

\item{genes_gr}{A GRanges object for genes. It must contain the information
for all the homolog ids.}

\item{symbol_colunm_name}{The column name of gene symbol in 'genes_gr'.}
}
\value{
A data frame with annotated homologs.
}
\description{
Add chromosome, start information to the homolog data.
}
