% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{buildChromDF}
\alias{buildChromDF}
\title{Build chromosome plotting dataframes}
\usage{
buildChromDF(chrom_infos, chr_orders)
}
\arguments{
\item{chrom_infos}{A list with the chromosome information data.frame.}

\item{chr_orders}{A list with the ordered chromosome names}
}
\value{
A list with the ordered chromosome plot information
}
\description{
Build chromosome plotting dataframes
}
