% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{buildHomologLinksDF}
\alias{buildHomologLinksDF}
\title{Build the homolog links dataframe}
\usage{
buildHomologLinksDF(homolog_df, chrom_df)
}
\arguments{
\item{homolog_df}{The data.frame with homologs. The output from function
\link{subsetHomologsByChrom}}

\item{chrom_df}{A list with the ordered chromosome plot information. The
output of function}
}
\value{
A list with the homolog links plot information.
}
\description{
Build the homolog links dataframe
}
