% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{filterChrom}
\alias{filterChrom}
\title{Filter chromosomes}
\usage{
filterChrom(chrInfo, sp_min_chr_size = 1e+07)
}
\arguments{
\item{chrInfo}{A data.frame retrieved from Ensembl by function
getChromInfoFromEnsembl.}

\item{sp_min_chr_size}{The minimal size for a chromosome.}
}
\description{
Filter chromosomes
}
