% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{get_unique_max_rows}
\alias{get_unique_max_rows}
\title{Assign each column to its best row, ensuring no duplicate rows
Greedy approach: process columns by their max value (highest first)}
\usage{
get_unique_max_rows(count_matrix)
}
\arguments{
\item{count_matrix}{the count table.}
}
\description{
Assign each column to its best row, ensuring no duplicate rows
Greedy approach: process columns by their max value (highest first)
}
