% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getGeneGRs}
\alias{getGeneGRs}
\title{Retrieve all the gene ranges}
\usage{
getGeneGRs(ids, marts, homologs)
}
\arguments{
\item{ids}{A named list with the gene ids for each species}

\item{marts}{A named list with the Mart object for each species}

\item{homologs}{A list of homologs. It must be the output from function
\link[orthoRibbon:getHomologGRs]{getHomologGRs}l}
}
\value{
A list with gene information GRanges.
}
\description{
Retrieve all the gene ranges
}
\examples{
if(interactive()){
 library(biomaRt)
 marts <- useEnsembl(
  biomart = "ENSEMBL_MART_ENSEMBL",
  dataset = "hsapiens_gene_ensembl",
  host='https://jun2026.archive.ensembl.org')
 common_name <- setNames('hsapiens', 'hsapiens')
 ids <- getGeneIDs(common_name, marts)
 homologs <- getHomologGRs(ids, common_name, marts)
 genes_gr <- getGeneGRs(ids, marts, homologs)
}

}
