% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getHomologGRs}
\alias{getHomologGRs}
\title{Retrieve the homolog pairs}
\usage{
getHomologGRs(ids, common_name, marts)
}
\arguments{
\item{ids}{A named list with the gene ids for each species}

\item{common_name}{species abbreviations e.g. "hsapiens", "mmusculus", "drerio"}

\item{marts}{A named list with the Mart object for each species}
}
\value{
A list with homolog GRanges.
}
\description{
Retrieve the homolog pairs
}
\examples{
\dontrun{
common_name <- setNames(nm=c('hsapiens', 'mmusculus'))
# This will retrieve data via bioMart
# prepare mart for BioMart
marts <- geneClusterPattern::guessSpecies(
  common_name, output='mart',
  host='https://jun2026.archive.ensembl.org')

# prepare all the Ensembl ids
ids <- orthoRibbon::getGeneIDs(common_name, marts)

# retrieve all the homologs
homologs <- orthoRibbon::getHomologGRs(ids, common_name, marts)
}

}
