% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomologPlotDataClass.R
\docType{class}
\name{HomologPlotData-class}
\alias{HomologPlotData-class}
\alias{HomologPlotData}
\alias{$,HomologPlotData-method}
\alias{$<-,HomologPlotData-method}
\alias{[[,HomologPlotData,ANY,ANY-method}
\alias{"[[",HomologPlotData,ANY,ANY-method}
\alias{[,HomologPlotData,ANY,ANY,ANY-method}
\alias{"[",HomologPlotData,ANY,ANY,ANY-method}
\alias{names,HomologPlotData-method}
\alias{"names",HomologPlotData-method}
\alias{.DollarNames.HomologPlotData}
\alias{.DollarNames,HomologPlotData,character-method}
\alias{subset,HomologPlotData-method}
\alias{subset,HomologPlotData,list-method}
\title{HomologPlotData S4 class}
\usage{
HomologPlotData(
  homolog_df_list,
  chrom_bars_df,
  chrom_label_df,
  symbol_df_top,
  symbol_df_bottom,
  common_name
)

\S4method{$}{HomologPlotData}(x, name)

\S4method{$}{HomologPlotData}(x, name) <- value

\S4method{[[}{HomologPlotData,ANY,ANY}(x, i, j, ...)

\S4method{[}{HomologPlotData,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{names}{HomologPlotData}(x)

\method{.DollarNames}{HomologPlotData}(x, pattern = "")

\S4method{subset}{HomologPlotData}(x, subset, ...)
}
\arguments{
\item{homolog_df_list}{list of data.frame with top and bottom plot information.}

\item{chrom_bars_df}{data.frame of chromosome plot information.}

\item{chrom_label_df}{data.frame of chromosome labels information.}

\item{symbol_df_top, symbol_df_bottom}{data.frame of symbols for top and
bottom layer.}

\item{common_name}{species abbreviations e.g. "hsapiens", "mmusculus", "drerio"}

\item{x}{A HomologPlotData object.}

\item{name, i}{slot name of HomologPlotData.}

\item{value}{value to be set}

\item{j, drop, ...}{not used.}

\item{pattern}{A regular expression. Only matching names are returned.}

\item{subset}{is a named list, e.g.
list(speciesA = c("chr1", "chr2"), speciesB = c("chr1", "chrX"))
Names must match values in chrom_bars_df$sp; each element's values must
match either chrom_bars_df$chrom or chrom_bars_df$label for that species.
The first name is treated as the "top" species, the last as the "bottom"
species, and homolog_df_list / symbol_df_top / symbol_df_bottom are
filtered to match.}
}
\value{
An object of HomologPlotData.
}
\description{
Wraps the collection of data frames/lists used for an orthoRibbon plot into
a single S4 object, with list-like \code{$}, \code{[[}, and \code{[} access.
}
\section{Slots}{

\describe{
\item{\code{homolog_df_list}}{list of data.frame with top and bottom plot information.}

\item{\code{chrom_bars_df}}{data.frame of chromosome plot information.}

\item{\code{chrom_label_df}}{data.frame of chromosome labels information.}

\item{\code{symbol_df_top,symbol_df_bottom}}{data.frame of symbols for top and
bottom layer.}

\item{\code{common_name}}{species abbreviations e.g. "hsapiens", "mmusculus", "drerio"}
}}

\examples{
HomologPlotData(
  homolog_df_list=list(),
  chrom_bars_df=data.frame(),
  chrom_label_df=data.frame(),
  symbol_df_top=data.frame(),
  symbol_df_bottom=data.frame(),
  common_name=setNames(nm=c('hsapiens', 'mmusculus'))
)

}
