% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{subsetHomologsByChrom}
\alias{subsetHomologsByChrom}
\title{Filter the homolog data.frame by chromosome names}
\usage{
subsetHomologsByChrom(homolog_df, chrom_infos, max_links = 5000)
}
\arguments{
\item{homolog_df}{The data.frame with homologs. The output from function
\link{addGeneInfo}}

\item{chrom_infos}{A list with the chromosome information data.frame.}

\item{max_links}{A numeric. The maximal link number to show in the plot.}
}
\value{
A data.frame with filtered homologs
}
\description{
Filter the homolog data.frame by chromosome names
}
