% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{addGeneInfo}
\alias{addGeneInfo}
\title{Add gene information to homolog data.frame}
\usage{
addGeneInfo(homolog_df, genes_gr)
}
\arguments{
\item{homolog_df}{The data.frame with homologs. The output from function
\link{getHomologIDs}}

\item{genes_gr}{The GRanges object with gene info. The output from function
\link{getGeneGRs}}
}
\description{
Add gene information to homolog data.frame
}
