% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{buildBezierDF}
\alias{buildBezierDF}
\title{Build the data.frame for Bezier curve}
\usage{
buildBezierDF(
  homolog_df_list,
  colname1,
  colname2,
  col1 = "seq_top",
  col2 = "seq_bottom",
  cl1 = 4
)
}
\arguments{
\item{homolog_df_list}{A list with the data.frame of plot data for homologs.}

\item{colname1, colname2}{The column names for top and bottom final positions.}

\item{col1, col2}{The column names for color.}

\item{cl1}{A nummeric. The Bezier curve is created with two control points.
And cl1 should be no more than 4 and no less than 0.
If it is set to 4, all points in Bezier curve will be set to col1.
If it is set to 3, the top point, two control points will be set to col1 and
the bottom point will be set to col2.
If it is set to 2, the top point, top control point will be set to col1 and
the bottom control point, bottom point will be set to col2.
If it is set to 1, the top point will be set to col1 and others col2.
If it is set to 0, all points will be set to col2.}
}
\description{
Build the data.frame for Bezier curve
}
\examples{
# example code

}
