% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPlotData.R
\name{buildPlotData}
\alias{buildPlotData}
\title{Build the plot data}
\usage{
buildPlotData(
  common_name,
  homolog_df,
  genes_gr,
  chrom_infos,
  sp_min_chr_size = 1e+07,
  chr_orders = NULL,
  filterByCoordSystem = TRUE,
  max_links = Inf,
  chromosome_order_method = "max"
)
}
\arguments{
\item{common_name}{Species abbreviations e.g. "hsapiens", "mmusculus", "drerio"}

\item{homolog_df}{The data.frame for homologs with column names
"gene_id1", "gene_id2".}

\item{genes_gr}{A GRanges object for genes. It must contain the information
for all the homolog ids.
'gene_name', 'species' must be metadata column name of 'genes_gr'.}

\item{chrom_infos}{A list of chromosome information.
The Chromosome information for each species must be a data.frame with columns
"name" and "length".}

\item{sp_min_chr_size}{The minimal chromosome size.}

\item{chr_orders}{Chromosome orders for plot.}

\item{filterByCoordSystem}{Filter the chromosomes or not.}

\item{max_links}{A numeric. The maximal link number to show in the plot.}

\item{chromosome_order_method}{A character string with the name of the
seriation method or 'spearman' or 'max'.
If using 'max', it will try to find the best order by maximal number of
homologous for each chromosome pairs.
If using 'spearman', it will try to find the best order by the Spearman
distance.
Otherwise see \link[seriation:seriate]{seriate}. It will be worth to
try 'TSP' first.}
}
\value{
A list with elements "homolog_df_list", "chrom_bars_df",
"chrom_label_df" , "symbol_df_top" and, "symbol_df_bottom" for plot.
}
\description{
Build the plot data from homolog data.frame, gene annotations and chromosome
informations.
}
\examples{
# example code

}
