% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getGeneGRs}
\alias{getGeneGRs}
\title{Retrieve all the gene ranges}
\usage{
getGeneGRs(ids, marts, homologs)
}
\arguments{
\item{ids}{A named list with the gene ids for each species}

\item{marts}{A named list with the Mart object for each species}

\item{homologs}{A list of homologs. It must be the output from function
\link[orthoRibbon:getHomologGRs]{getHomologGRs}l}
}
\value{
A list with gene information GRanges.
}
\description{
Retrieve all the gene ranges
}
