% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getGeneIDs}
\alias{getGeneIDs}
\title{Retrieve all the ensembl gene IDs by given species abbreviations}
\usage{
getGeneIDs(com_name, marts)
}
\arguments{
\item{com_name}{species abbreviations eg. "hsapiens", "mmusculus", "drerio"}

\item{marts}{A named list with the Mart object for each species}
}
\value{
A list of character with ensembl gene IDs.
}
\description{
Retrieve all the ensembl gene IDs by given species abbreviations
}
