% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getHomologGRs}
\alias{getHomologGRs}
\title{Retrieve the homolog pairs}
\usage{
getHomologGRs(ids, com_name, marts)
}
\arguments{
\item{ids}{A named list with the gene ids for each species}

\item{com_name}{species abbreviations eg. "hsapiens", "mmusculus", "drerio"}

\item{marts}{A named list with the Mart object for each species}
}
\value{
A list with homolog GRanges.
}
\description{
Retrieve the homolog pairs
}
