% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getHomologIDs}
\alias{getHomologIDs}
\title{Retrieve the homolog pairs}
\usage{
getHomologIDs(homologs)
}
\arguments{
\item{homologs}{A named list with the homolog GRanges. The output from
function \link{getHomologGRs}.}
}
\value{
A data.frame with homologs.
}
\description{
Retrieve the homolog pairs
}
