% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoRibbonPlot.R
\name{orthoRibbonPlot}
\alias{orthoRibbonPlot}
\title{Plot the data}
\usage{
orthoRibbonPlot(
  common_name,
  bezier_df,
  chrom_bars_df,
  chrom_label_df,
  symbol_df_top,
  symbol_df_bottom,
  link_lwd = 0.25,
  link_alpha = 1,
  chr_lwd = 3,
  chr_lineend = "round",
  chr_size = 6,
  label_size = 12,
  symbol_size = 2,
  xlim = c(-0.02, 1.02),
  show_symbol = FALSE
)
}
\arguments{
\item{common_name}{Species abbreviations e.g. "hsapiens", "mmusculus", "drerio".}

\item{bezier_df}{A data.frame. The coordinates for Bezier curves.
It must have columns "x", "y", "id", and "col".
The "id" column saves group id for each pair of homologs.}

\item{chrom_bars_df}{A data.frame. The coordinates for chromosome line.
It must have columns "x", "y", "sp" and "chrom". Here "sp" means species and
"chrom" indicates the chromosome.}

\item{chrom_label_df}{A data.frame. The coordinates for chromosome names.
It must have columns "x", "y" and "labels"}

\item{symbol_df_top, symbol_df_bottom}{The data.frame for the gene labels}

\item{link_lwd, chr_lwd}{The line width for Bezier curve, chromosome bar.}

\item{link_alpha}{The alpha value for Bezier curve.}

\item{chr_lineend}{The line end for chromosome bar.}

\item{chr_size, label_size, symbol_size}{The size for chromosome label,
species, and symbols.}

\item{xlim}{Two numeric values, specifying the x limit of the scale.}

\item{show_symbol}{A logical value. Show the symbols or not.}
}
\value{
A ggplot object
}
\description{
Plot data by ribbon plot.
}
\examples{
common_name <- setNames(nm=c("hsapiens", "dmelanogaster"))
extdata <- system.file('extdata', 'human_fly', package='orthoRibbon')
chrom_infos <- readRDS(file.path(extdata, 'chrom_infos.rds'))
homologs_df <- readRDS(file.path(extdata, 'homologs_df.rds'))
plotData <- buildPlotData(common_name, homologs_df,
                          chrom_infos = chrom_infos,
                          chromosome_order_method = 'max',
                          max_links=1000)
bezier_chr <- buildBezierDF(plotData$homolog_df_list,
                            colname1='topChr_finalOffset',
                            colname2='bottomChr_finalOffset')
orthoRibbonPlot(common_name, bezier_chr,
                plotData$chrom_bars_df,
                plotData$chrom_label_df)
}
