% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoRibbonPlot.R
\name{orthoRibbonPlot}
\alias{orthoRibbonPlot}
\title{Plot the data}
\usage{
orthoRibbonPlot(
  com_name,
  bezier_df,
  chrom_bars_df,
  chrom_label_df,
  symbol_list_top,
  symbol_list_bottom,
  link_lwd = 0.25,
  chr_lwd = 3,
  chr_lineend = "round",
  chr_size = 6,
  label_size = 12,
  symbol_size = 2,
  xlim = c(-0.02, 1.02),
  show_symbol = FALSE
)
}
\arguments{
\item{com_name}{Species abbreviations e.g. "hsapiens", "mmusculus", "drerio".}

\item{bezier_df}{A data.frame. The coordinates for Bezier curves.
It must have columns "x", "y", "id", and "col".
The "id" column saves group id for each pair of homologs.}

\item{chrom_bars_df}{A data.frame. The coordinates for chromosome line.
It must have columns "x", "y", "sp" and "chrom". Here "sp" means species and
"chrom" indicates the chromosome.}

\item{chrom_label_df}{A data.frame. The coordinates for chromosome names.
It must have columns "x", "y" and "labels"}

\item{symbol_list_top, symbol_list_bottom}{The data.frame for the gene labels}

\item{link_lwd, chr_lwd}{The line width for Bezier curve, chromosome bar,}

\item{chr_lineend}{The line end for chromosome bar.}

\item{chr_size, label_size, symbol_size}{The size for chromosome label,
species, and symbols.}

\item{xlim}{Two numeric values, specifying the x limit of the scale.}

\item{show_symbol}{A logical value. Show the symbols or not.}
}
\value{
A ggplot object
}
\description{
Plot data by ribbon plot.
}
\examples{
# example code

}
